agent/requirements.txt
langchain
langchain-groq
python-dotenv
x402
eth_account
